Move oauth tests to their own files.
Add tests for duplicate fitbit and withings notifications.
Add test scenarios for login with existing users.
Move oauth tests to routes folder.
