+++ tests/routes/test_withings_routes.py
@@ -2,17 +2,13 @@ import dataclasses
 import datetime
 import json
 import math
-import re
 
 import pytest
-from authlib.integrations.starlette_client.apps import StarletteOAuth2App
 from fastapi import status
-from fastapi.responses import RedirectResponse
 from fastapi.testclient import TestClient
 from httpx import Response
 from respx import MockRouter
 
-from slackhealthbot.data.database.connection import ctx_db
 from slackhealthbot.data.database.models import User
 from slackhealthbot.data.database.models import WithingsUser as DbWithingsUser
 from slackhealthbot.data.repositories import withingsrepository
@@ -128,55 +124,35 @@ async def test_weight_notification(
 
 
 @pytest.mark.asyncio
-async def test_refresh_token_ok(
+async def test_duplicate_weight_notification(
     mocked_async_session,
     client: TestClient,
     respx_mock: MockRouter,
     withings_factories: tuple[UserFactory, WithingsUserFactory],
 ):
     """
-    Given a user whose access token is expired
-    When we receive the callback from withings that a new weight is available
-    Then the access token is refreshed
-    And the latest weight is updated in the database
-    And the message is posted to slack with the correct pattern.
+    Given a user with a given previous weight logged
+    When we receive the callback twice from withings that a new weight is available
+    Then the last_weight is updated in the database
+    And the message is posted to slack only once
     """
+
     user_factory, withings_user_factory = withings_factories
 
-    ctx_db.set(mocked_async_session)
     # Given a user
     user: User = user_factory(withings=None)
     db_withings_user: DbWithingsUser = withings_user_factory(
         user_id=user.id,
         last_weight=50.2,
-        oauth_access_token="some old access token",
         oauth_expiration_date=datetime.datetime.now(datetime.timezone.utc)
-        - datetime.timedelta(days=1),
-    )
-
-    # Mock withings oauth refresh token success
-    oauth_token_refresh_request = respx_mock.post(
-        url=f"{settings.withings_base_url}v2/oauth2",
-    ).mock(
-        Response(
-            status_code=200,
-            json={
-                "status": 0,
-                "body": {
-                    "userid": user.withings.oauth_userid,
-                    "access_token": "some new access token",
-                    "refresh_token": "some new refresh token",
-                    "expires_in": 600,
-                },
-            },
-        )
+        + datetime.timedelta(days=1),
     )
 
     # Mock withings endpoint to return some weight data
-    withings_weight_request = respx_mock.post(
+    weight_request = respx_mock.post(
         url=f"{settings.withings_base_url}measure",
     ).mock(
-        Response(
+        return_value=Response(
             status_code=200,
             json={
                 "status": 0,
@@ -193,7 +169,7 @@ async def test_refresh_token_ok(
                     ],
                 },
             },
-        ),
+        )
     )
 
     # Mock an empty ok response from the slack webhook
@@ -202,8 +178,10 @@ async def test_refresh_token_ok(
     )
 
     # When we receive the callback from withings that a new weight is available
-    with client as client_ctx:
-        response = client_ctx.post(
+    # Use the client as a context manager so the app can have its lfespan events triggered.
+    # https://fastapi.tiangolo.com/advanced/testing-events/
+    with client:
+        response = client.post(
             "/withings-notification-webhook/",
             data={
                 "userid": db_withings_user.oauth_userid,
@@ -214,72 +192,24 @@ async def test_refresh_token_ok(
 
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
-    repo_user: withingsrepository.User = (
-        await withingsrepository.get_user_by_withings_userid(
+    # Then the last_weight is updated in the database
+    assert weight_request.call_count == 1
+    fitness_data: withingsrepository.FitnessData = (
+        await withingsrepository.get_fitness_data_by_withings_userid(
             mocked_async_session,
             withings_userid=db_withings_user.oauth_userid,
         )
     )
-    # Then the access token is refreshed.
-    assert withings_weight_request.call_count == 1
-    assert (
-        withings_weight_request.calls[0].request.headers["authorization"]
-        == "Bearer some new access token"
-    )
-    assert oauth_token_refresh_request.call_count == 1
-    assert repo_user.oauth_data.oauth_access_token == "some new access token"
-    assert repo_user.oauth_data.oauth_refresh_token == "some new refresh token"
-
-    # And the last_weight is updated in the database
-    assert math.isclose(repo_user.fitness_data.last_weight_kg, 50.05)
+    assert math.isclose(fitness_data.last_weight_kg, 50.05)
 
     # And the message is sent to slack as expected
+    assert slack_request.call_count == 1
     actual_message = json.loads(slack_request.calls[0].request.content)["text"]
     assert "↘️" in actual_message
 
-
-@pytest.mark.asyncio
-async def test_refresh_token_fail(
-    mocked_async_session,
-    client: TestClient,
-    respx_mock: MockRouter,
-    withings_factories: tuple[UserFactory, WithingsUserFactory],
-):
-    """
-    Given a user whose access token is expired and invalid
-    When we receive the callback from withings that a new weight is available
-    Then the access token refresh fails
-    And no weight is updated in the database
-    And the message is posted to slack about the user being logged out
-    """
-    user_factory, withings_user_factory = withings_factories
-
-    ctx_db.set(mocked_async_session)
-    # Given a user
-    user: User = user_factory(withings=None, slack_alias="jdoe")
-    db_withings_user: DbWithingsUser = withings_user_factory(
-        user_id=user.id,
-        last_weight=None,
-        oauth_access_token="some old invalid access token",
-        oauth_expiration_date=datetime.datetime.now(datetime.timezone.utc)
-        - datetime.timedelta(days=1),
-    )
-
-    # Mock withings oauth refresh token success
-    oauth_token_refresh_request = respx_mock.post(
-        url=f"{settings.withings_base_url}v2/oauth2",
-    ).mock(
-        Response(status_code=200, json={"status": 100})  # TODO confirm code
-    )
-
-    # Mock an empty ok response from the slack webhook
-    slack_request = respx_mock.post(url=f"{settings.slack_webhook_url}").mock(
-        return_value=Response(status_code=200)
-    )
-
-    # When we receive the callback from withings that a new weight is available
-    with client as client_ctx:
-        response = client_ctx.post(
+    # When we receive the callback a second time from withings
+    with client:
+        response = client.post(
             "/withings-notification-webhook/",
             data={
                 "userid": db_withings_user.oauth_userid,
@@ -289,184 +219,6 @@ async def test_refresh_token_fail(
         )
 
     assert response.status_code == status.HTTP_204_NO_CONTENT
-
-    repo_user: withingsrepository.User = (
-        await withingsrepository.get_user_by_withings_userid(
-            mocked_async_session,
-            withings_userid=db_withings_user.oauth_userid,
-        )
-    )
-    # Then the access token is not refreshed.
-    assert oauth_token_refresh_request.call_count == 1
-    assert repo_user.oauth_data.oauth_access_token == "some old invalid access token"
-
-    # And no new weight data is updated in the database
-    repo_fitness_data: withingsrepository.FitnessData = (
-        await withingsrepository.get_fitness_data_by_withings_userid(
-            db=mocked_async_session,
-            withings_userid=repo_user.identity.withings_userid,
-        )
-    )
-    assert repo_fitness_data.last_weight_kg is None
-
-    # And a message was sent to slack about the user being logged out
-    actual_message = json.loads(slack_request.calls[0].request.content)["text"].replace(
-        "\n", ""
-    )
-    assert re.search(
-        "Oh no <@jdoe>, looks like you were logged out of withings! 😳.", actual_message
-    )
-
-
-@pytest.mark.asyncio
-async def test_login_success(
-    mocked_async_session,
-    client: TestClient,
-    monkeypatch: pytest.MonkeyPatch,
-):
-    ctx_db.set(mocked_async_session)
-
-    # mock authlib's generation of a url on withings
-    async def mock_authorize_redirect(fake_self, request, redirect_uri):
-        return RedirectResponse("https://fakewithings.com", status_code=302)
-
-    monkeypatch.setattr(
-        StarletteOAuth2App,
-        "authorize_redirect",
-        mock_authorize_redirect,
-    )
-    # Simulate the user starting the withings login
-    with client:
-        response: Response = client.get(
-            "/v1/withings-authorization/jdoe",
-            follow_redirects=False,
-        )
-    assert response.status_code == status.HTTP_302_FOUND
-    assert response.headers["location"] == "https://fakewithings.com"
-
-    # mock authlib's token response
-    async def mock_authorize_access_token(fake_self, request):
-        return {
-            "userid": "user123",
-            "access_token": "some access token",
-            "refresh_token": "some refresh token",
-            "expires_in": 3600,
-        }
-
-    monkeypatch.setattr(
-        StarletteOAuth2App,
-        "authorize_access_token",
-        mock_authorize_access_token,
-    )
-
-    # Simulate withings's response to the sleep and activity subscriptions
-    async def mock_post(fake_self, *args, **kwargs):
-        return Response(status_code=200, json={"status": 0})
-
-    monkeypatch.setattr(
-        StarletteOAuth2App,
-        "post",
-        mock_post,
-    )
-
-    # Simulate withings calling us back to finish the login
-    with client:
-        response: Response = client.get(
-            "/withings-oauth-webhook",
-        )
-
-    assert response.status_code == status.HTTP_200_OK
-
-    # Verify that we have the expected data in the db
-    repo_user = await withingsrepository.get_user_by_withings_userid(
-        mocked_async_session,
-        "user123",
-    )
-    assert repo_user.identity == withingsrepository.UserIdentity(
-        withings_userid="user123",
-        slack_alias="jdoe",
-    )
-
-    assert repo_user.oauth_data.oauth_access_token == "some access token"
-    assert repo_user.oauth_data.oauth_refresh_token == "some refresh token"
-
-
-@pytest.mark.asyncio
-async def test_logged_out(
-    mocked_async_session,
-    client: TestClient,
-    respx_mock: MockRouter,
-    withings_factories: tuple[UserFactory, WithingsUserFactory],
-):
-    """
-    Given a user whose access token is invalid
-    When we receive the callback from withings that a new weight is available
-    Then no weight is updated in the database
-    And a message is posted to slack about the user being logged out
-    """
-
-    user_factory, withings_user_factory = withings_factories
-
-    ctx_db.set(mocked_async_session)
-    # Given a user
-    user: User = user_factory(withings=None, slack_alias="jdoe")
-    db_withings_user: DbWithingsUser = withings_user_factory(
-        user_id=user.id,
-        last_weight=None,
-        oauth_access_token="some invalid access token",
-        oauth_expiration_date=datetime.datetime.now(datetime.timezone.utc)
-        + datetime.timedelta(days=1),
-    )
-
-    # Mock withings endpoint to return an unauthorized error
-    withings_weight_request = respx_mock.post(
-        url=f"{settings.withings_base_url}measure",
-    ).mock(
-        return_value=Response(
-            status_code=200, json={"status": 100}  # TODO confirm code
-        )
-    )
-
-    # Mock an empty ok response from the slack webhook
-    slack_request = respx_mock.post(f"{settings.slack_webhook_url}").mock(
-        return_value=Response(200)
-    )
-    # When we receive the callback from withings that a new weight is available
-    with client as client_ctx:
-        response = client_ctx.post(
-            "/withings-notification-webhook/",
-            data={
-                "userid": db_withings_user.oauth_userid,
-                "startdate": 1683894606,
-                "enddate": 1686570821,
-            },
-        )
-
-    assert response.status_code == status.HTTP_204_NO_CONTENT
-    repo_user: withingsrepository.User = (
-        await withingsrepository.get_user_by_withings_userid(
-            mocked_async_session,
-            withings_userid=db_withings_user.oauth_userid,
-        )
-    )
-
-    # Then the access token is not refreshed.
-    assert withings_weight_request.call_count == 1
-    assert repo_user.oauth_data.oauth_access_token == "some invalid access token"
-
-    # And no new weight data is updated in the database
-    repo_fitness_data: withingsrepository.FitnessData = (
-        await withingsrepository.get_fitness_data_by_withings_userid(
-            db=mocked_async_session,
-            withings_userid=repo_user.identity.withings_userid,
-        )
-    )
-    assert repo_fitness_data.last_weight_kg is None
-
-    # And a message was sent to slack about the user being logged out
-    actual_message = json.loads(slack_request.calls[0].request.content)["text"].replace(
-        "\n", ""
-    )
-    assert re.search(
-        "Oh no <@jdoe>, looks like you were logged out of withings! 😳.", actual_message
-    )
+    # Then we don't post to stack a second time
+    assert weight_request.call_count == 1
+    assert slack_request.call_count == 1
